clean up length prefixing
