+++ src/Binrep/Type/Prefix/Size.hs
@@ -1,0 +1,56 @@
+{-# LANGUAGE AllowAmbiguousTypes #-}
+{-# LANGUAGE UndecidableInstances #-} -- required for easier instances
+{-# LANGUAGE OverloadedStrings #-} -- for refined errors
+
+{- TODO 2023-02-20
+  * split into sizeprefix (Type) and lengthprefix (Type -> Type)
+  * better behaviour for lengthprefix, sizeprefix is a special case
+-}
+
+module Binrep.Type.Prefix.Size where
+
+import Binrep.Type.Prefix
+import Binrep
+import FlatParse.Basic qualified as FP
+
+import GHC.TypeNats
+import Util.TypeNats ( natValInt )
+import Data.ByteString qualified as B
+import Refined hiding ( Weaken(..), strengthen )
+import Refined.Unsafe
+
+import Data.Typeable ( Typeable, typeRep )
+
+data SizePrefix pfx
+type SizePrefixed pfx = Refined (SizePrefix pfx)
+
+instance (KnownNat (Max pfx), BLen a, Typeable pfx)
+  => Predicate (SizePrefix pfx) a where
+    validate p a
+      | blen a <= natValInt @(Max pfx) = Nothing
+      | otherwise = throwRefineOtherException (typeRep p) $
+          "thing too big for length prefix type"
+
+-- TODO no idea if this is sensible
+instance IsCBLen (SizePrefixed pfx a) where
+    type CBLen (SizePrefixed pfx a) = CBLen pfx + CBLen a
+
+instance (Prefix pfx, BLen a, BLen pfx)
+  => BLen (SizePrefixed pfx a) where
+    blen ra = blen (lenToPfx @pfx (blen a)) + blen a
+      where a = unrefine ra
+
+instance (Prefix pfx, BLen a, Put pfx, Put a)
+  => Put (SizePrefixed pfx a) where
+    put ra = put (lenToPfx @pfx (blen a)) <> put a
+      where a = unrefine ra
+
+class GetSize a where getSize :: Int -> Getter a
+instance GetSize B.ByteString where getSize = FP.take
+
+instance (Prefix pfx, GetSize a, Get pfx)
+  => Get (SizePrefixed pfx a) where
+    get = do
+        pfx <- get @pfx
+        a <- getSize (pfxToLen pfx)
+        pure $ reallyUnsafeRefine a
